OCSADV-200: moved ProgramGen to pot bundle
